requirements.txt
# Telegram Bot with job queue support
python-telegram-bot[job-queue]==20.3

# 环境变量
python-dotenv==1.0.0

# HTTP请求
requests==2.31.0

# WebSocket
websocket-client==1.6.1

# 工具库
pytz==2023.3
python-dateutil==2.8.2

# Job Queue依赖
APScheduler==3.10.1
